+++ src/main/java/com/yuuko/core/commands/animal/commands/DogCommand.java
@@ -11,13 +11,15 @@ import net.dv8tion.jda.core.EmbedBuilder;
 import java.util.Arrays;
 public class DogCommand extends Command {
 
+    private static final String BASE_URL = "https://dog.ceo/api/breeds/image/random";
+
     public DogCommand() {
         super("dog", AnimalModule.class, 0, Arrays.asList("-dog"), false, null);
     }
 
     @Override
     public void onCommand(MessageEvent e) {
-        final JsonObject object = new RequestHandler("https://dog.ceo/api/breeds/image/random").getJsonObject();
+        final JsonObject object = new RequestHandler(BASE_URL).getJsonObject();
 
         EmbedBuilder embed = new EmbedBuilder().setTitle("Random Dog")
                 .setImage(object.get("message").getAsString());