+++ src/main/java/com/yuuko/core/commands/media/commands/LondonUndergroundCommand.java
@@ -23,6 +23,8 @@ import java.util.List;
 
 public class LondonUndergroundCommand extends Command {
 
+    private static final String BASE_URL = "https://api.tfl.gov.uk/line/mode/tube/status?app_id=";
+
     public LondonUndergroundCommand() {
         super("underground", MediaModule.class, 0, Arrays.asList("-underground", "-underground <min>"), false, null);
     }
@@ -30,7 +32,7 @@ public class LondonUndergroundCommand extends Command {
     @Override
     public void onCommand(MessageEvent e) {
         try {
-            final String url = "https://api.tfl.gov.uk/line/mode/tube/status?app_id=" + Utilities.getApiApplicationId("transportforlondon") + "&app_key=" + Utilities.getApiKey("transportforlondon");
+            final String url = BASE_URL + Utilities.getApiApplicationId("transportforlondon") + "&app_key=" + Utilities.getApiKey("transportforlondon");
             final String json = new RequestHandler(url).getString();
 
             ArrayList<LineManager> lineManager = new ObjectMapper().readValue(json, new TypeReference<List<LineManager>>(){});