+++ src/main/java/com/yuuko/core/commands/media/commands/TescoCommand.java
@@ -8,6 +8,7 @@ import com.yuuko.core.commands.media.MediaModule;
 import com.yuuko.core.events.entity.MessageEvent;
 import com.yuuko.core.io.RequestHandler;
 import com.yuuko.core.io.entity.RequestProperty;
+import com.yuuko.core.utilities.Sanitiser;
 import com.yuuko.core.utilities.Utilities;
 import net.dv8tion.jda.core.EmbedBuilder;
 
@@ -17,6 +18,8 @@ import java.util.Arrays;
 
 public class TescoCommand extends Command {
 
+    private static final String BASE_URL = "https://dev.tescolabs.com/grocery/products/?query=";
+
     public TescoCommand() {
         super("tesco", MediaModule.class, 1, Arrays.asList("-tesco <product>"), false, null);
     }
@@ -24,7 +27,7 @@ public class TescoCommand extends Command {
     @Override
     public void onCommand(MessageEvent e) {
         try {
-            final String url = "https://dev.tescolabs.com/grocery/products/?query=" + e.getParameters().replace(" ", "%20") + "&offset=0&limit=1";
+            final String url = BASE_URL + Sanitiser.scrubString(e.getParameters(), true) + "&offset=0&limit=1";
             final JsonObject json = new RequestHandler(url, new RequestProperty("Ocp-Apim-Subscription-Key", Utilities.getApiKey("tesco"))).getJsonObject();
             final JsonObject preData = json.get("uk").getAsJsonObject().get("ghs").getAsJsonObject().get("products").getAsJsonObject();