+++ src/main/java/com/yuuko/core/commands/media/commands/WeatherCommand.java
@@ -7,6 +7,7 @@ import com.yuuko.core.commands.Command;
 import com.yuuko.core.commands.media.MediaModule;
 import com.yuuko.core.events.entity.MessageEvent;
 import com.yuuko.core.io.RequestHandler;
+import com.yuuko.core.utilities.Sanitiser;
 import com.yuuko.core.utilities.Utilities;
 import net.dv8tion.jda.core.EmbedBuilder;
 
@@ -18,6 +19,8 @@ import java.util.Arrays;
 
 public class WeatherCommand extends Command {
 
+    private static final String BASE_URL = "https://api.openweathermap.org/data/2.5/weather?q=";
+
     public WeatherCommand() {
         super("weather", MediaModule.class, 1, Arrays.asList("-weather <city>", "-weather <city> <country>"), false, null);
     }
@@ -25,7 +28,7 @@ public class WeatherCommand extends Command {
     @Override
     public void onCommand(MessageEvent e) {
         try {
-            final String url = "https://api.openweathermap.org/data/2.5/weather?q=" + (e.getParameters().replace(" ", "+")) + "&units=metric&APPID=" + Utilities.getApiKey("openweathermap");
+            final String url = BASE_URL + (Sanitiser.scrubString(e.getParameters(), false).replace(" ", "+")) + "&units=metric&APPID=" + Utilities.getApiKey("openweathermap");
             JsonObject data = new RequestHandler(url).getJsonObject();
 
             if(data == null) {