+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
-## 2019-07-09
+## 2019-07-13
++ [+] Added new method to the Sanitiser class to scrub strings of special characters.
++ [~] Added static BASE_URL field to most commands where a URL is used every execute.
 + [~] Improved guarding of the [__nuke__] command.
 + [~] Moved [__urban__] command from nsfw module to media since it's a better fit, although it might not bit the final home.